agente1-indexador/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
PyMuPDF
python-multipart
